pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vibe"
version = "0.1.0"
description = "VIBE — Vacuum-Interaction Birefringence Explorer"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "numpy",
  "scipy",
  "matplotlib",
  "pandas",
  "opencv-python",
  "tqdm",
  "pyyaml",
  "ipympl",
  "astropy",
  "LightPipes",
  "h5py",
  "Pillow",
  "scikit-image",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
